REF move virtual package tests
